+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## _**v2.9.4**_ (09-20-2023)
+
+- [FIX] Fixed issues with some sensors not updating there status.
+- [NEW] Added support for Air Quality sensors.
+
 ## _**v2.9.3**_ (09-20-2023)
 
 - [FIX] Testing some fixes for the Garage Door Status and contacts